+++ packages/assets/src/scss/inputs/_checkbox.scss
@@ -1,6 +1,29 @@
 @use '../functions' as *;
 @use '../variables' as *;
 
+@mixin checkbox-active-state {
+    border-color: $color-primary-80;
+    background-color: $color-primary-80;
+
+    &::after {
+        border-color: $color-neutral-10;
+    }
+
+    &:disabled {
+        border-color: $color-primary-60;
+        background-color: $color-primary-60;
+    }
+
+    &.ids-input--error {
+        border-color: $color-error-80;
+        background-color: $color-error-80;
+
+        &::after {
+            border-color: $color-neutral-10;
+        }
+    }
+}
+
 .ids-checkbox {
     .ids-input--checkbox {
         -webkit-appearance: none;
@@ -27,43 +50,8 @@
             transform: rotate(-45deg);
         }
 
-        &:checked,
-        &:indeterminate {
-            border-color: $color-primary-80;
-            background-color: $color-primary-80;
-
-            &::after {
-                border-color: $color-neutral-10;
-            }
-
-            &:disabled {
-                border-color: $color-primary-60;
-                background-color: $color-primary-60;
-            }
-
-            &.ids-input--error {
-                border-color: $color-error-80;
-                background-color: $color-error-80;
-
-                &::after {
-                    border-color: $color-neutral-10;
-                }
-            }
-        }
-
-        &:indeterminate {
-            &::after {
-                background-color: $color-neutral-10;
-                height: calculateRem(2px);
-                top: 50%;
-                transform: translateY(-50%);
-            }
-
-            &.ids-input--error {
-                &::after {
-                    background-color: $color-neutral-10;
-                }
-            }
+        &:checked {
+            @include checkbox-active-state;
         }
     }
 }
