+++ packages/assets/src/scss/styles.scss
@@ -14,5 +14,6 @@
 
 @use 'inputs/checkbox';
 @use 'inputs/input-text';
+@use 'inputs/three-state-checkbox';
 
 @use 'ui/clear-btn';
