+++ packages/components/src/inputs/Checkbox/Checkbox.stories.tsx
@@ -16,7 +16,7 @@ const meta: Meta<typeof CheckboxStateful> = {
         title: {
             control: 'text',
         },
-        value: {
+        checked: {
             control: 'boolean',
         },
     },
@@ -59,7 +59,7 @@ export const Checked: Story = {
     name: 'Checked',
     args: {
         name: 'default-input',
-        value: true,
+        checked: true,
     },
 };
 
@@ -68,7 +68,7 @@ export const CheckedDisabled: Story = {
     args: {
         name: 'default-input',
         disabled: true,
-        value: true,
+        checked: true,
     },
 };
 
@@ -77,6 +77,6 @@ export const CheckedError: Story = {
     args: {
         name: 'default-input',
         error: true,
-        value: true,
+        checked: true,
     },
 };
