+++ packages/components/src/internal/partials/BaseCheckbox/index.ts
@@ -1,0 +1,5 @@
+import BaseCheckbox from './BaseCheckbox';
+import { BaseCheckboxProps } from './BaseCheckbox.types';
+
+export default BaseCheckbox;
+export type { BaseCheckboxProps };
